Moved function to RequestHelpers
